Add a fast Create Plan page - step 6
Changes:
1. use plain map for resolving subpaths
2. share common routines for storage and network mappings
3. refactorings
  a) extract types to separate file
  b) move reducer related files into a folder
  c) replace toSpliced() -  unsupported ECMA 2023 feature
4. add target provider to the dependency list for hooks that take it
   as parameter. Fixes a bug where providers pointing to the same
   physical env do not trigger useEffect().
5. provided messages for basic error states
6. validate that the target namespace is not already used by
   one of the selected VMs
7. disable/pause queries that depend from params fetched dynamically
   until that data is fetched. Fixes problem with partial state in case
   of missing params.
