+++ packages/forklift-console-plugin/locales/en/plugin__forklift-console-plugin.json
@@ -28,6 +28,9 @@
   "Actions": "Actions",
   "Add mapping": "Add mapping",
   "Add source and target providers for the migration.": "Add source and target providers for the migration.",
+  "All discovered networks have been mapped to the default network.": "All discovered networks have been mapped to the default network.",
+  "All discovered storages have been mapped to the default storage.": "All discovered storages have been mapped to the default storage.",
+  "API Error": "API Error",
   "Application credential ID": "Application credential ID",
   "Application credential name": "Application credential name",
   "Application credential secret": "Application credential secret",
@@ -235,12 +238,15 @@
   "Namespace is not defined": "Namespace is not defined",
   "Network for data transfer": "Network for data transfer",
   "Network interfaces": "Network interfaces",
+  "Network Map name re-generated": "Network Map name re-generated",
   "Network map:": "Network map:",
+  "Network mappings have been re-generated": "Network mappings have been re-generated",
   "NetworkAttachmentDefinitions": "NetworkAttachmentDefinitions",
   "NetworkMaps": "NetworkMaps",
   "NetworkMaps for virtualization": "NetworkMaps for virtualization",
   "Networks": "Networks",
   "Networks used by the selected VMs": "Networks used by the selected VMs",
+  "New name was generated for the Network Map due to naming conflict.": "New name was generated for the Network Map due to naming conflict.",
   "No credentials found.": "No credentials found.",
   "No inventory data available.": "No inventory data available.",
   "No NetworkMaps found in namespace <1>{namespace}</1>.": "No NetworkMaps found in namespace <1>{namespace}</1>.",
@@ -249,6 +255,7 @@
   "No owner": "No owner",
   "No Plans found in namespace <1>{namespace}</1>.": "No Plans found in namespace <1>{namespace}</1>.",
   "No Plans found.": "No Plans found.",
+  "No provider is available": "No provider is available",
   "No providers found": "No providers found",
   "No Providers found in namespace <1>{namespace}</1>.": "No Providers found in namespace <1>{namespace}</1>.",
   "No results found": "No results found",
@@ -259,6 +266,7 @@
   "No StorageMaps found in namespace <1>{namespace}</1>.": "No StorageMaps found in namespace <1>{namespace}</1>.",
   "No StorageMaps found.": "No StorageMaps found.",
   "No storages in this category": "No storages in this category",
+  "No target provider exists ": "No target provider exists ",
   "Not Ready": "Not Ready",
   "Note: If <1>Skip certificate validation</1> is selected, migrations from this provider will not be secure, meaning that the transferred data is sent over an insecure connection and potentially sensitive data could be exposed.": "Note: If <1>Skip certificate validation</1> is selected, migrations from this provider will not be secure, meaning that the transferred data is sent over an insecure connection and potentially sensitive data could be exposed.",
   "Note: If this field is checked/true, migrations from this provider will not be secure, meaning that the transferred data is sent over an insecure connection and potentially sensitive data could be exposed.": "Note: If this field is checked/true, migrations from this provider will not be secure, meaning that the transferred data is sent over an insecure connection and potentially sensitive data could be exposed.",
@@ -347,6 +355,8 @@
   "Secret failed to load, check if the secret exist.": "Secret failed to load, check if the secret exist.",
   "Secret is loading, please wait.": "Secret is loading, please wait.",
   "Secrets": "Secrets",
+  "Select a namespace": "Select a namespace",
+  "Select a provider": "Select a provider",
   "Select migration network": "Select migration network",
   "Select migration target": "Select migration target",
   "Select provider type": "Select provider type",
@@ -377,6 +387,7 @@
   "Storage classes": "Storage classes",
   "Storage domains": "Storage domains",
   "Storage map:": "Storage map:",
+  "Storage mappings have been re-generated": "Storage mappings have been re-generated",
   "StorageMaps": "StorageMaps",
   "StorageMaps for virtualization": "StorageMaps for virtualization",
   "Storages used by the selected VMs": "Storages used by the selected VMs",
@@ -386,6 +397,7 @@
   "Target provider": "Target provider",
   "Template": "Template",
   "Tenant": "Tenant",
+  "The chosen provider is no longer available.": "The chosen provider is no longer available.",
   "The interval in minutes for precopy. Default value is 60.": "The interval in minutes for precopy. Default value is 60.",
   "The interval in seconds for snapshot pooling. Default value is 10.": "The interval in seconds for snapshot pooling. Default value is 10.",
   "The limit for CPU usage by the controller, specified in milliCPU. Default value is 500m.": "The limit for CPU usage by the controller, specified in milliCPU. Default value is 500m.",
