+++ packages/forklift-console-plugin/package.json
@@ -26,10 +26,13 @@
     "@kubev2v/common": "*",
     "@kubev2v/legacy": "*",
     "@migtools/lib-ui": "8.4.1",
+    "immer": "^10.0.3",
     "jsonpath": "^1.1.1",
     "luxon": "^3.3.0",
     "msw": "^1.2.3",
-    "streamsaver": "^2.0.6"
+    "streamsaver": "^2.0.6",
+    "use-immer": "^0.9.0",
+    "uuid": "^9.0.1"
   },
   "peerDependencies": {
     "@openshift-console/dynamic-plugin-sdk": "0.0.18",
@@ -52,7 +55,6 @@
     "@types/jsonpath": "^0.2.0",
     "copy-webpack-plugin": "^11.0.0",
     "css-loader": "^6.7.1",
-    "immer": "^10.0.3",
     "mini-svg-data-uri": "^1.4.4",
     "react-linkify": "^1.0.0-alpha",
     "style-loader": "^3.3.1",
@@ -60,8 +62,6 @@
     "ts-loader": "^9.3.1",
     "tsconfig-paths-webpack-plugin": "^4.0.0",
     "typescript": "^4.7.4",
-    "use-immer": "^0.9.0",
-    "uuid": "^9.0.1",
     "webpack": "^5.79.0",
     "webpack-cli": "^5.0.2",
     "webpack-dev-server": "^4.15.1"
