+++ packages/forklift-console-plugin/src/modules/Providers/hooks/useNamespaces.ts
@@ -1,3 +1,4 @@
+import { useMemo } from 'react';
 import { useProviderInventory } from 'src/modules/Providers/hooks';
 
 import { OpenShiftNamespace, V1beta1Provider } from '@kubev2v/types';
@@ -12,7 +13,10 @@ export const useNamespaces = (
   } = useProviderInventory<OpenShiftNamespace[]>({
     provider,
     subPath: 'namespaces',
+    disabled: !provider,
   });
 
-  return [!loading && !error && Array.isArray(namespaces) ? namespaces : [], loading, error];
+  const stableResponse = useMemo(() => (Array.isArray(namespaces) ? namespaces : []), [namespaces]);
+
+  return [stableResponse, loading, error];
 };
