+++ packages/forklift-console-plugin/src/modules/Providers/views/migrate/reducer/getNamespacesUsedBySelectedVms.ts
@@ -1,0 +1,10 @@
+import { VmData } from '../../details';
+
+export const getNamespacesUsedBySelectedVms = (selectedVMs: VmData[]) =>
+  Array.from(
+    new Set(
+      selectedVMs
+        .map((d) => (d?.vm?.providerType === 'openshift' ? d.vm.namespace : undefined))
+        .filter(Boolean),
+    ),
+  );
