+++ packages/forklift-console-plugin/src/modules/Providers/views/migrate/reducer/getStoragesUsedBySelectedVMs.ts
@@ -1,6 +1,6 @@
 import { OpenstackVolume, OVirtDisk } from '@kubev2v/types';
 
-import { VmData } from '../details';
+import { VmData } from '../../details';
 
 // based on packages/legacy/src/Plans/components/Wizard/helpers.tsx
 export const getStoragesUsedBySelectedVms = (
